requirements.txt
numpy>=1.23
pandas>=2.0
scipy>=1.10
pyyaml>=6.0
tqdm>=4.66
torch>=2.1
torchvision>=0.16
pillow>=10.0
opencv-python>=4.8
matplotlib>=3.8
scikit-learn>=1.3
jupyter>=1.0
ipykernel>=6.25